+++ lib/teleprompter.dart
@@ -3,4 +3,5 @@ library teleprompter;
 export 'package:teleprompter/src/data/capture_mode.dart';
 export 'package:teleprompter/src/teleprompter_widget.dart';
 export 'package:teleprompter/src/ui/camera/recording_marker.dart';
+export 'package:teleprompter/src/ui/camera/overlay_camera_widget.dart';
 export 'package:teleprompter/src/ui/timer/stopwatch_widget.dart';